Move join request-related views into sub-module
