+++ coldfront/core/project/views.py
@@ -1,12 +1,10 @@
 import datetime
 
-from django.conf import settings
 from django.contrib import messages
 from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
 from django.contrib.auth.models import User
 from django.contrib.messages.views import SuccessMessageMixin
 from django.core.paginator import EmptyPage, PageNotAnInteger, Paginator
-from django.db import transaction
 from django.db.models import Case, CharField, Q, Value, When
 from django.forms import formset_factory
 from django.http import HttpResponse, HttpResponseRedirect
@@ -31,26 +29,18 @@ from coldfront.core.project.forms import (ProjectAddUserForm,
                                           ProjectAddUsersToAllocationForm,
                                           ProjectReviewEmailForm,
                                           ProjectReviewForm,
-                                          ProjectReviewUserJoinForm,
                                           ProjectSearchForm,
                                           ProjectUpdateForm,
-                                          ProjectUserUpdateForm,
-                                          JoinRequestSearchForm,
-                                          ProjectSelectHostUserForm)
+                                          ProjectUserUpdateForm)
 from coldfront.core.project.models import (Project, ProjectReview,
                                            ProjectReviewStatusChoice,
                                            ProjectStatusChoice, ProjectUser,
-                                           ProjectUserJoinRequest,
                                            ProjectUserRoleChoice,
                                            ProjectUserStatusChoice,
                                            ProjectUserRemovalRequest,
-                                           ProjectUserRemovalRequestStatusChoice,
-                                           SavioProjectAllocationRequest,
-                                           VectorProjectAllocationRequest)
+                                           ProjectUserRemovalRequestStatusChoice)
 from coldfront.core.project.utils import (annotate_queryset_with_cluster_name,
-                                          is_primary_cluster_project,
-                                          send_project_join_notification_email,
-                                          send_project_join_request_denial_email)
+                                          is_primary_cluster_project)
 from coldfront.core.project.utils_.addition_utils import can_project_purchase_service_units
 from coldfront.core.project.utils_.new_project_user_utils import NewProjectUserRunnerFactory
 from coldfront.core.project.utils_.new_project_user_utils import NewProjectUserSource
@@ -59,7 +49,7 @@ from coldfront.core.project.utils_.renewal_utils import is_any_project_pi_renewa
 from coldfront.core.resource.utils_.allowance_utils.computing_allowance import ComputingAllowance
 from coldfront.core.resource.utils_.allowance_utils.interface import ComputingAllowanceInterface
 from coldfront.core.user.forms import UserSearchForm
-from coldfront.core.user.utils import CombinedUserSearch, needs_host
+from coldfront.core.user.utils import CombinedUserSearch
 from coldfront.core.utils.common import (get_domain_url, import_from_settings)
 from coldfront.core.utils.mail import send_email, send_email_template
 
@@ -1335,592 +1325,3 @@ class ProjectReivewEmailView(LoginRequiredMixin, UserPassesTestMixin, FormView):
 
     def get_success_url(self):
         return reverse('project-review-list')
-
-
-class ProjectJoinView(LoginRequiredMixin, UserPassesTestMixin, TemplateView):
-    login_url = '/'
-
-    logger = logging.getLogger(__name__)
-
-    def test_func(self):
-        project_obj = get_object_or_404(Project, pk=self.kwargs.get('pk'))
-        user_obj = self.request.user
-        project_users = project_obj.projectuser_set.filter(user=user_obj)
-        reason = self.request.POST.get('reason')
-
-        if self.request.user.userprofile.access_agreement_signed_date is None:
-            messages.error(
-                self.request, 'You must sign the User Access Agreement before you can join a project.')
-            return False
-
-        inactive_project_status = ProjectStatusChoice.objects.get(
-            name='Inactive')
-        if project_obj.status == inactive_project_status:
-            message = (
-                f'Project {project_obj.name} is inactive, and may not be '
-                f'joined.')
-            messages.error(self.request, message)
-            return False
-
-        if project_users.exists():
-            project_user = project_users.first()
-            if project_user.status.name == 'Active':
-                message = (
-                    f'You are already a member of Project {project_obj.name}.')
-                messages.error(self.request, message)
-                return False
-            if project_user.status.name == 'Pending - Add':
-                message = (
-                    f'You have already requested to join Project '
-                    f'{project_obj.name}.')
-                messages.warning(self.request, message)
-                return False
-
-            pending_status = ProjectUserRemovalRequestStatusChoice.objects.get(name='Pending')
-            processing_status = ProjectUserRemovalRequestStatusChoice.objects.get(name='Processing')
-
-            if ProjectUserRemovalRequest.objects. \
-                    filter(project_user=project_user,
-                           status__in=[pending_status, processing_status]).exists():
-                message = (
-                    f'You cannot join Project {project_obj.name} because you '
-                    f'have a pending removal request for '
-                    f'{project_obj.name}.')
-                messages.error(self.request, message)
-                return False
-
-        # If the user is the requester or PI on a pending request for the
-        # Project, do not allow the join request.
-        if project_obj.name.startswith('vector_'):
-            request_model = VectorProjectAllocationRequest
-        else:
-            request_model = SavioProjectAllocationRequest
-        is_requester_or_pi = Q(requester=user_obj) | Q(pi=user_obj)
-        if request_model.objects.filter(
-                is_requester_or_pi, project=project_obj,
-                status__name__in=['Under Review', 'Approved - Processing']):
-            message = (
-                f'You are the requester or PI of a pending request for '
-                f'Project {project_obj.name}, so you may not join it. You '
-                f'will automatically be added when it is approved.')
-            messages.warning(self.request, message)
-            return False
-
-        if len(reason) < 20:
-            message = 'Please provide a valid reason to join the project (min 20 characters)'
-            messages.error(self.request, message)
-            return False
-
-        return True
-
-    def get(self, *args, **kwargs):
-        return redirect(self.login_url)
-
-    def post(self, request, *args, **kwargs):
-        project_obj = get_object_or_404(Project, pk=self.kwargs.get('pk'))
-        user_obj = self.request.user
-        project_users = project_obj.projectuser_set.filter(user=user_obj)
-        role = ProjectUserRoleChoice.objects.get(name='User')
-        status = ProjectUserStatusChoice.objects.get(name='Pending - Add')
-        reason = self.request.POST['reason']
-
-        select_host_user_form = ProjectSelectHostUserForm(
-            project=project_obj.name,
-            data=self.request.POST)
-        host_user = None
-        if select_host_user_form.is_valid():
-            host_user = \
-                User.objects.get(
-                    username=select_host_user_form.cleaned_data['host_user'])
-
-        if project_users.exists():
-            project_user = project_users.first()
-            project_user.role = role
-            # If the user is Active on the project, raise a warning and exit.
-            if project_user.status.name == 'Active':
-                message = (
-                    f'You are already an Active member of Project '
-                    f'{project_obj.name}.')
-                messages.warning(self.request, message)
-                next_view = reverse('project-join-list')
-                return redirect(next_view)
-            project_user.status = status
-            project_user.save()
-        else:
-            project_user = ProjectUser.objects.create(
-                user=user_obj,
-                project=project_obj,
-                role=role,
-                status=status)
-
-        # Create a join request
-        ProjectUserJoinRequest.objects.create(project_user=project_user,
-                                              reason=reason,
-                                              host_user=host_user)
-
-        message = (
-            f'You have requested to join Project {project_obj.name}. The '
-            f'managers have been notified.')
-        messages.success(self.request, message)
-        next_view = reverse('project-join-list')
-
-        # Send a notification to the project managers.
-        try:
-            send_project_join_notification_email(project_obj, project_user)
-        except Exception as e:
-            message = 'Failed to send notification email. Details:'
-            self.logger.error(message)
-            self.logger.exception(e)
-
-        return redirect(next_view)
-
-
-class ProjectJoinListView(ProjectListView, UserPassesTestMixin):
-
-    template_name = 'project/project_join_list.html'
-
-    def test_func(self):
-        user = self.request.user
-        if user.userprofile.access_agreement_signed_date is None:
-            message = (
-                'You must sign the User Access Agreement before you can join '
-                'a project.')
-            messages.error(self.request, message)
-            return False
-        return True
-
-    def get_queryset(self):
-
-        order_by = self.request.GET.get('order_by')
-        if order_by:
-            direction = self.request.GET.get('direction')
-            if direction == 'asc':
-                direction = ''
-            else:
-                direction = '-'
-            order_by = direction + order_by
-        else:
-            order_by = 'id'
-
-        project_search_form = ProjectSearchForm(self.request.GET)
-
-        projects = Project.objects.prefetch_related(
-            'field_of_science', 'status').filter(
-                status__name__in=['New', 'Active', ]
-        ).order_by(order_by)
-        projects = annotate_queryset_with_cluster_name(projects)
-
-        if project_search_form.is_valid():
-            data = project_search_form.cleaned_data
-
-            # Last Name
-            if data.get('last_name'):
-                pi_project_users = ProjectUser.objects.filter(
-                    project__in=projects,
-                    role__name='Principal Investigator',
-                    user__last_name__icontains=data.get('last_name'))
-                project_ids = pi_project_users.values_list(
-                    'project_id', flat=True)
-                projects = projects.filter(id__in=project_ids)
-
-            # Username
-            if data.get('username'):
-                projects = projects.filter(
-                    Q(projectuser__user__username__icontains=data.get(
-                        'username')) &
-                    (Q(projectuser__role__name='Principal Investigator') |
-                     Q(projectuser__status__name='Active'))
-                )
-
-            # Field of Science
-            if data.get('field_of_science'):
-                projects = projects.filter(
-                    field_of_science__description__icontains=data.get(
-                        'field_of_science'))
-
-            # Project Title
-            if data.get('project_title'):
-                projects = projects.filter(title__icontains=data.get('project_title'))
-
-            # Project Name
-            if data.get('project_name'):
-                projects = projects.filter(name__icontains=data.get('project_name'))
-
-            # Cluster Name
-            if data.get('cluster_name'):
-                projects = projects.filter(cluster_name__icontains=data.get('cluster_name'))
-
-        return projects.distinct()
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        projects = self.get_queryset()
-        user_obj = self.request.user
-
-        # A User may not join a Project he/she is already a pending or active
-        # member of.
-        already_pending_or_active = set(projects.filter(
-            projectuser__user=user_obj,
-            projectuser__status__name__in=['Pending - Add', 'Active', ]
-        ).values_list('name', flat=True))
-        # A User may not join a Project with a pending
-        # SavioProjectAllocationRequest where he/she is the requester or PI.
-        is_requester_or_pi = Q(requester=user_obj) | Q(pi=user_obj)
-        pending_project_request_statuses = [
-            'Under Review', 'Approved - Processing']
-        is_part_of_pending_savio_project_request = set(
-            SavioProjectAllocationRequest.objects.prefetch_related(
-                'project'
-            ).filter(
-                is_requester_or_pi,
-                status__name__in=pending_project_request_statuses
-            ).values_list('project__name', flat=True))
-        # A User may not join a Project with a pending
-        # VectorProjectAllocationRequest where he/she is the requester or PI.
-        is_part_of_pending_vector_project_request = set(
-            VectorProjectAllocationRequest.objects.prefetch_related(
-                'project'
-            ).filter(
-                is_requester_or_pi,
-                status__name__in=pending_project_request_statuses
-            ).values_list('project__name', flat=True))
-        pending_removal_requests = set([removal_request.project_user.project.name
-                                        for removal_request in
-                                        ProjectUserRemovalRequest.objects.filter(
-                                            Q(project_user__user__username=self.request.user.username) &
-                                            Q(status__name='Pending'))])
-
-        not_joinable = set.union(
-            already_pending_or_active,
-            is_part_of_pending_savio_project_request,
-            is_part_of_pending_vector_project_request,
-            pending_removal_requests)
-
-        join_requests = Project.objects.filter(Q(projectuser__user=self.request.user)
-                                               & Q(status__name__in=['New', 'Active', ])
-                                               & Q(projectuser__status__name__in=['Pending - Add']))
-        join_requests = annotate_queryset_with_cluster_name(join_requests)
-
-        context['join_requests'] = join_requests
-        context['not_joinable'] = not_joinable
-
-        # Only non-LBL employees without a host user and without any pending
-        # join requests need access to the SelectHostUserForm.
-        context['need_host'] = False
-        pending_status = ProjectUserStatusChoice.objects.get(name='Pending - Add')
-        if flag_enabled('LRC_ONLY') \
-                and needs_host(self.request.user) \
-                and not ProjectUser.objects.filter(user=self.request.user,
-                                                   status=pending_status).exists():
-            context['need_host'] = True
-
-            selecthostform_dict = {}
-            for project in context.get('project_list'):
-                selecthostform_dict[project.name] = \
-                    ProjectSelectHostUserForm(project=project.name)
-
-            context['selecthostform_dict'] = selecthostform_dict
-
-        return context
-
-
-class ProjectReviewJoinRequestsView(LoginRequiredMixin, UserPassesTestMixin,
-                                    TemplateView):
-    form_class = ProjectReviewUserJoinForm
-    template_name = 'project/project_review_join_requests.html'
-
-    project_obj = None
-    redirect = None
-    users_to_review = []
-
-    def test_func(self):
-        return (
-            self._is_superuser_or_project_owner() or
-            self.request.user.has_perm('project.can_view_all_projects'))
-
-    def dispatch(self, request, *args, **kwargs):
-        self.project_obj = get_object_or_404(
-            Project.objects.prefetch_related('projectuser_set'),
-            pk=self.kwargs.get('pk'))
-        self.redirect = HttpResponseRedirect(
-            reverse('project-detail', kwargs={'pk': self.project_obj.pk}))
-        if self.project_obj.status.name not in ['Active', 'New', ]:
-            message = 'You cannot review join requests to an archived project.'
-            messages.error(request, message)
-            return self.redirect
-        self.users_to_review = self._get_users_to_review()
-        return super().dispatch(request, *args, **kwargs)
-
-    def get(self, request, *args, **kwargs):
-        context = {}
-
-        users_to_review = self.users_to_review
-        if users_to_review:
-            formset = formset_factory(
-                self.form_class, max_num=len(users_to_review))
-            formset = formset(initial=users_to_review, prefix='userform')
-            context['formset'] = formset
-
-        context['project'] = self.project_obj
-
-        context['can_add_users'] = self._is_superuser_or_project_owner()
-
-        if flag_enabled('LRC_ONLY'):
-            context['host_dict'] = self._get_host_dict()
-
-        return render(request, self.template_name, context)
-
-    def post(self, request, *args, **kwargs):
-        # Some users who have GET privileges do not have POST privileges.
-        if not self._is_superuser_or_project_owner():
-            message = 'You do not have permission to access this page.'
-            messages.error(request, message)
-            return self.redirect
-
-        users_to_review = self.users_to_review
-        formset = formset_factory(
-            self.form_class, max_num=len(users_to_review))
-        formset = formset(
-            request.POST, initial=users_to_review, prefix='userform')
-
-        if not formset.is_valid():
-            for error in formset.errors:
-                messages.error(request, error)
-            return self.redirect
-
-        decision = request.POST.get('decision', None)
-        if decision == 'approve':
-            status_name = 'Active'
-            message_verb = 'Approved'
-        elif decision == 'deny':
-            status_name = 'Denied'
-            message_verb = 'Denied'
-        else:
-            message = f'Unexpected review decision: {decision}.'
-            messages.error(request, message)
-            return self.redirect
-
-        project_user_status_choice = ProjectUserStatusChoice.objects.get(
-            name=status_name)
-
-        num_reviews = 0
-        failed_usernames = []
-        for form in formset:
-            user_form_data = form.cleaned_data
-            if not user_form_data['selected']:
-                continue
-            username = user_form_data.get('username')
-            user_obj = User.objects.get(username=username)
-            project_user_obj = self.project_obj.projectuser_set.get(
-                user=user_obj)
-            try:
-                self._process_project_user_request(
-                    project_user_obj, project_user_status_choice)
-            except Exception as e:
-                logger.exception(e)
-                failed_usernames.append(username)
-            else:
-                num_reviews += 1
-
-        num_failures = len(failed_usernames)
-        num_successes = num_reviews - num_failures
-
-        message = (
-            f'{message_verb} {num_successes}/{num_reviews} user requests to '
-            f'join the project. {settings.PROGRAM_NAME_SHORT} staff have been '
-            f'notified to set up cluster access for each approved request.')
-        if num_failures > 0:
-            messages.warning(request, message)
-            failed_usernames_str = ', '.join(failed_usernames)
-            message = f'Failed to process requests by: {failed_usernames_str}.'
-            messages.error(request, message)
-        else:
-            messages.success(request, message)
-
-        return self.redirect
-
-    def _get_host_dict(self):
-        """Return a mapping from username to the associated host user,
-        which may be None, for each user to review."""
-        host_dict = {}
-        for user in self.users_to_review:
-            username = user.get('username')
-            join_requests = ProjectUserJoinRequest.objects.filter(
-                project_user__project=self.project_obj,
-                project_user__user__username=username,
-                host_user__isnull=False)
-            if join_requests.exists():
-                host_user = join_requests.latest('modified').host_user
-            else:
-                host_user = None
-            host_dict[username] = host_user
-        return host_dict
-
-    def _get_users_to_review(self):
-        """Return a list of dictionaries representing Users who have
-        made requests to join the Project."""
-        users_to_review = []
-        queryset = self.project_obj.projectuser_set.filter(
-            status__name='Pending - Add').order_by('user__username')
-        for ele in queryset:
-            try:
-                reason = ele.projectuserjoinrequest_set.latest(
-                    'created').reason
-            except ProjectUserJoinRequest.DoesNotExist:
-                reason = ProjectUserJoinRequest.DEFAULT_REASON
-
-            user = {
-                'username': ele.user.username,
-                'first_name': ele.user.first_name,
-                'last_name': ele.user.last_name,
-                'email': ele.user.email,
-                'role': ele.role,
-                'reason': reason
-            }
-            users_to_review.append(user)
-        return users_to_review
-
-    def _is_superuser_or_project_owner(self):
-        """Return whether the requesting user is a superuser or an
-        'Active' Project PI or manager."""
-        user = self.request.user
-        if user.is_superuser:
-            return True
-        if self.project_obj.projectuser_set.filter(
-                user=user,
-                role__name__in=['Manager', 'Principal Investigator'],
-                status__name='Active').exists():
-            return True
-        return False
-
-    @staticmethod
-    def _process_project_user_request(project_user_obj,
-                                      project_user_status_choice):
-        """Given a ProjectUser, set its status to the given one, and run
-        any additional processing."""
-        with transaction.atomic():
-            project_user_obj.status = project_user_status_choice
-            project_user_obj.save()
-            if project_user_status_choice.name == 'Active':
-                runner_factory = NewProjectUserRunnerFactory()
-                runner = runner_factory.get_runner(
-                    project_user_obj, NewProjectUserSource.JOINED)
-                runner.run()
-            else:
-                try:
-                    send_project_join_request_denial_email(
-                        project_user_obj.project, project_user_obj)
-                except Exception as e:
-                    message = (
-                        f'Failed to send notification email. Details:\n{e}')
-                    logger.exception(message)
-
-
-class ProjectJoinRequestListView(LoginRequiredMixin, UserPassesTestMixin,
-                                 ListView):
-    template_name = 'project/project_join_request_list.html'
-    paginate_by = 25
-
-    def get_queryset(self):
-        order_by = self.request.GET.get('order_by')
-        if order_by:
-            direction = self.request.GET.get('direction')
-            if direction == 'asc':
-                direction = ''
-            else:
-                direction = '-'
-            order_by = direction + 'created'
-        else:
-            order_by = '-created'
-
-        project_join_requests = \
-            ProjectUserJoinRequest.objects.filter(
-                pk__in=ProjectUserJoinRequest.objects.filter(
-                    project_user__status__name=
-                    'Pending - Add').order_by(
-                    'project_user', '-created').distinct(
-                    'project_user'))
-
-        join_request_search_form = JoinRequestSearchForm(self.request.GET)
-
-        if join_request_search_form.is_valid():
-            data = join_request_search_form.cleaned_data
-
-            if data.get('username'):
-                project_join_requests = \
-                    project_join_requests.filter(
-                        project_user__user__username__icontains=data.get('username'))
-
-            if data.get('email'):
-                project_join_requests = \
-                    project_join_requests.filter(
-                        project_user__user__email__icontains=data.get('email'))
-
-            if data.get('project_name'):
-                project_join_requests = \
-                    project_join_requests.filter(
-                        project_user__project__name__icontains=data.get('project_name'))
-
-        return project_join_requests.order_by(order_by)
-
-    def test_func(self):
-        """UserPassesTestMixin tests."""
-        if self.request.user.is_superuser:
-            return True
-
-        if self.request.user.has_perm('project.view_projectuserjoinrequest'):
-            return True
-
-        message = (
-            'You do not have permission to view project join requests.')
-        messages.error(self.request, message)
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        join_request_search_form = JoinRequestSearchForm(self.request.GET)
-        if join_request_search_form.is_valid():
-            context['join_request_search_form'] = join_request_search_form
-            data = join_request_search_form.cleaned_data
-            filter_parameters = ''
-            for key, value in data.items():
-                if value:
-                    if isinstance(value, list):
-                        for ele in value:
-                            filter_parameters += '{}={}&'.format(key, ele)
-                    else:
-                        filter_parameters += '{}={}&'.format(key, value)
-            context['join_request_search_form'] = join_request_search_form
-        else:
-            filter_parameters = None
-            context['join_request_search_form'] = JoinRequestSearchForm()
-
-        order_by = self.request.GET.get('order_by')
-        if order_by:
-            direction = self.request.GET.get('direction')
-            filter_parameters_with_order_by = filter_parameters + \
-                                              'order_by=%s&direction=%s&' % (order_by, direction)
-        else:
-            filter_parameters_with_order_by = filter_parameters
-
-        context['expand_accordion'] = 'show'
-
-        context['filter_parameters'] = filter_parameters
-        context['filter_parameters_with_order_by'] = filter_parameters_with_order_by
-
-        join_request_queryset = self.get_queryset()
-
-        paginator = Paginator(join_request_queryset, self.paginate_by)
-
-        page = self.request.GET.get('page')
-
-        try:
-            join_requests = paginator.page(page)
-        except PageNotAnInteger:
-            join_requests = paginator.page(1)
-        except EmptyPage:
-            join_requests = paginator.page(paginator.num_pages)
-
-        context['join_request_list'] = join_requests
-
-        return context
